Avoid passing `null` to `parse_str`
If `->getQueryAsArray()` is called when no query string is present on the request, `parse_url()` here returns `null` instead of an empty string. Passing `null` to `parse_str()` is deprecated as of PHP 8.1. WordPress treats deprecations as errors with `WP_DEBUG` set to `true`, so fixing deprecations becomes necessary for better DX in this case.